Rename package to Nextcloud
On top of our old ownCloud package this package now provides Nextcloud.
It does not conflict with old ownCloud packaging and makes a smooth transition possible (with a helper package).

As Nextcloud did not rename the 'occ' commandline tool, the system-wide
installation has been moved to a seperate package nextcloud-occ which
uses alternatives to provide the /usr/bin/occ location from multiple
packages. Of course ownCloud packaging does not use this yet.